requirements.txt
# CVE Scanner Dashboard - Python Dependencies

# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1

# Optional: for better datetime handling
python-dateutil>=2.8.2

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
